Use new interface 'SGP'
There is no need to create an concatenated buffer of 'password:domain'
since all we do is to feed the bytes 1:1 through Md5. This changes
'func hashPassword(out []byte, password []byte)' to
'func hashPassword(out []byte, pw_parts ...[]byte)' and reduces the
amount of allocations.

Additionally, we also allocate a new temporary buffer in 'hashPassword'
on each call, just to throw it away at the end of the function.

'SGP' defines an interface which allows to implement a new method (sha1
comes to mind, see supergenpass.com). 'SGPMd5' implements 'SGP', reduces
the amount of allocations